updated main.py to launch GUI
